Move getApiUrl to sendMail
