+++ lib/emails/pin-code.js
@@ -1,5 +1,4 @@
 const {template} = require('lodash')
-const {getApiUrl} = require('./utils')
 
 const bodyTemplate = template(`
 <!DOCTYPE html>
@@ -46,7 +45,7 @@ const bodyTemplate = template(`
 
 <body>
   <div>
-    <img src="<%= apiUrl %>/public/images/logo-adresse.png" alt="Logo République Française">
+    <img src="$$API_URL$$/public/images/logo-adresse.png" alt="Logo République Française">
   </div>
   <div class="title">
     <h2 style="margin:0; mso-line-height-rule:exactly;">Publication de la Base Adresse Locale de <%= nomCommune %>
@@ -75,11 +74,10 @@ const bodyTemplate = template(`
 
 function formatEmail(data) {
   const {code, nomCommune} = data
-  const apiUrl = getApiUrl()
 
   return {
     subject: 'Demande de code d’identification',
-    html: bodyTemplate({code, nomCommune, apiUrl})
+    html: bodyTemplate({code, nomCommune})
   }
 }
