+++ lib/util/sendmail.js
@@ -1,6 +1,20 @@
 const nodemailer = require('nodemailer')
 const {pick} = require('lodash')
 
+function getApiUrl() {
+  if (process.env.API_DEPOT_URL) {
+    return process.env.API_DEPOT_URL
+  }
+
+  if (process.env.NODE_ENV === 'production') {
+    throw new Error('API_DEPOT_URL must be defined in production mode')
+  }
+
+  return 'https://plateforme.adresse.data.gouv.fr.local'
+}
+
+module.exports = {getApiUrl}
+
 function createTransport() {
   if (!process.env.SMTP_HOST && process.env.NODE_ENV === 'production') {
     throw new Error('SMTP_HOST must be provided in production mode')
@@ -31,6 +45,9 @@ async function sendMail(email, recipients = []) {
   if (recipients.length === 0) {
     throw new Error('At least one recipient must be provided')
   }
+
+  const apiUrl = getApiUrl()
+  email.html.replace('$$API_URL$$', apiUrl)
 
   const info = await transport.sendMail({
     ...pick(email, 'text', 'html', 'subject'),
